Hitsplat duration -> double; Quick KeyFrame routes by sprite
Two related Hitsplat-card tweaks:

1) HitsplatKeyFrame.duration int -> double. The DEFAULT_DURATION
   constant was already double (5/3 ticks) but the field, ctor param,
   and spinner were ints -- the spinner forced whole-tick rounding,
   so users could never pick the in-game 1.667-tick default exactly.
   Spinner model is now (0.1 increments) and createKeyFrame +
   AttributePanel cast through ((Number) val).doubleValue(). Existing
   readers (HitsplatOverlay, SummarySheet, AttributeSheet) already
   assigned getDuration() to local doubles and compared against -1
   so they didn't need to change.

2) Quick KeyFrame Hitsplat/Bar (renamed from Hitsplat/Health) now
   routes the damage-subtraction keyframe by sprite:
   - SHIELD / SHIELD_OTHER / SHIELD_MAX   -> Shield bar
   - POISE  / POISE_OTHER  / POISE_MAX    -> Special bar
   - everything else                      -> Health bar (legacy)
   For each bar, reads the previous keyframe of that bar type as the
   source of truth for duration / colour / max / current / order /
   width and creates a new keyframe at the seeker tick with current
   reduced by damage (clamped to 0). Falls back to the bar's natural
   defaults if no previous keyframe exists.

   Button tooltip explains the routing.

+++ src/main/java/com/creatorskit/swing/timesheet/keyframe/HitsplatKeyFrame.java
@@ -11,12 +11,19 @@ public class HitsplatKeyFrame extends KeyFrame
 {
     public static final double DEFAULT_DURATION = (double) 5 / 3;
 
-    private int duration;
+    /**
+     * Game ticks the hitsplat stays on screen. Fractional values allowed
+     * (0.1-tick increments in the spinner) so users can match the in-game
+     * 5/3-tick default exactly, plus shorter / longer effects. Sentinel -1
+     * means "use {@link #DEFAULT_DURATION}" -- see HitsplatOverlay /
+     * SummarySheet / AttributeSheet for the substitution.
+     */
+    private double duration;
     private HitsplatSprite sprite;
     private HitsplatVariant variant;
     private int damage;
 
-    public HitsplatKeyFrame(double tick, KeyFrameType hitsplatType, int duration, HitsplatSprite sprite, HitsplatVariant variant, int damage)
+    public HitsplatKeyFrame(double tick, KeyFrameType hitsplatType, double duration, HitsplatSprite sprite, HitsplatVariant variant, int damage)
     {
         super(hitsplatType, tick);
         this.duration = duration;
